move Indexer to util, set __all__
