+++ src/fast_forward/util/indexer.py
