+++ tests/test_indexer.py
@@ -4,7 +4,7 @@ import numpy as np
 
 from fast_forward.encoder import LambdaEncoder
 from fast_forward.index.memory import InMemoryIndex
-from fast_forward.indexer import Indexer
+from fast_forward.util.indexer import Indexer
 from fast_forward.quantizer.nanopq import NanoPQ
